sista

+++ src/realisering15/AgentAndraRasPaAlien.java
@@ -53,6 +53,7 @@ public class AgentAndraRasPaAlien extends javax.swing.JFrame {
         lblRasAttribut = new javax.swing.JLabel();
         btnBekrafta = new javax.swing.JButton();
         btnTillbaka = new javax.swing.JButton();
+        lblWorm = new javax.swing.JLabel();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
@@ -90,6 +91,8 @@ public class AgentAndraRasPaAlien extends javax.swing.JFrame {
             }
         });
 
+        lblWorm.setText("Om worm skriv 0");
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
@@ -103,12 +106,14 @@ public class AgentAndraRasPaAlien extends javax.swing.JFrame {
                         .addGap(32, 32, 32)
                         .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                             .addComponent(lblAlienID)
-                            .addGroup(layout.createSequentialGroup()
+                            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
                                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                                     .addComponent(cbValjAlienID, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                                     .addComponent(cbValjRas, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 183, Short.MAX_VALUE)
-                                .addComponent(txtRasAttribut, javax.swing.GroupLayout.PREFERRED_SIZE, 111, javax.swing.GroupLayout.PREFERRED_SIZE)))))
+                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                    .addComponent(lblWorm, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(txtRasAttribut, javax.swing.GroupLayout.DEFAULT_SIZE, 111, Short.MAX_VALUE))))))
                 .addGap(103, 103, 103))
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
@@ -146,7 +151,9 @@ public class AgentAndraRasPaAlien extends javax.swing.JFrame {
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(cbValjRas, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(txtRasAttribut, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 51, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(lblWorm)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 29, Short.MAX_VALUE)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(btnBekrafta)
                     .addComponent(btnTillbaka))
@@ -186,45 +193,46 @@ public class AgentAndraRasPaAlien extends javax.swing.JFrame {
 
 
     private void btnBekraftaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnBekraftaActionPerformed
-        try {
-            String valtID = cbValjAlienID.getSelectedItem().toString();
-            String valdRas = cbValjRas.getSelectedItem().toString();
-            String attribut = txtRasAttribut.getText();
+        if (Validering.textFaltHarVarde(txtRasAttribut) && Validering.isHeltal(txtRasAttribut) && Validering.isPositivt(txtRasAttribut)) {
+            try {
+                String valtID = cbValjAlienID.getSelectedItem().toString();
+                String valdRas = cbValjRas.getSelectedItem().toString();
+                String attribut = txtRasAttribut.getText();
 
-            //Tar bort från gamla rastabellen
-            ArrayList<String> bogloditer = databas.fetchColumn("SELECT alien_id FROM boglodite");
-            ArrayList<String> worms = databas.fetchColumn("SELECT alien_id FROM worm");
-            ArrayList<String> squids = databas.fetchColumn("SELECT alien_id FROM squid");
-            for (String enBoglodite : bogloditer) {
-                if (enBoglodite.equals(valtID)) {
-                    databas.delete("DELETE FROM boglodite WHERE alien_id=" + valtID);
+                //Tar bort från gamla rastabellen
+                ArrayList<String> bogloditer = databas.fetchColumn("SELECT alien_id FROM boglodite");
+                ArrayList<String> worms = databas.fetchColumn("SELECT alien_id FROM worm");
+                ArrayList<String> squids = databas.fetchColumn("SELECT alien_id FROM squid");
+                for (String enBoglodite : bogloditer) {
+                    if (enBoglodite.equals(valtID)) {
+                        databas.delete("DELETE FROM boglodite WHERE alien_id=" + valtID);
+                    }
                 }
-            }
-            for (String enWorm : worms) {
-                if (enWorm.equals(valtID)) {
-                    databas.delete("DELETE FROM worm WHERE alien_id=" + valtID);
+                for (String enWorm : worms) {
+                    if (enWorm.equals(valtID)) {
+                        databas.delete("DELETE FROM worm WHERE alien_id=" + valtID);
+                    }
                 }
-            }
-            for (String enSquid : squids) {
-                if (enSquid.equals(valtID)) {
-                    databas.delete("DELETE FROM squid WHERE alien_id=" + valtID);
+                for (String enSquid : squids) {
+                    if (enSquid.equals(valtID)) {
+                        databas.delete("DELETE FROM squid WHERE alien_id=" + valtID);
+                    }
                 }
-            }
 
-            //Uppdaterar ny ras och rasattribut
-            if (valdRas.equals("Boglodite")) {
-                databas.insert("INSERT INTO boglodite VALUES (" + valtID + ", " + attribut + ")");
-            } else if (valdRas.equals("Worm")) {
-                databas.insert("INSERT INTO worm VALUES (" + valtID + ")");
-            } else if (valdRas.equals("Squid")) {
-                databas.insert("INSERT INTO squid VALUES (" + valtID + ", " + attribut + ")");
-            }
-            JOptionPane.showMessageDialog(null, "Aliens ras och attribut har uppdaterats");
+                //Uppdaterar ny ras och rasattribut
+                if (valdRas.equals("Boglodite")) {
+                    databas.insert("INSERT INTO boglodite VALUES (" + valtID + ", " + attribut + ")");
+                } else if (valdRas.equals("Worm")) {
+                    databas.insert("INSERT INTO worm VALUES (" + valtID + ")");
+                } else if (valdRas.equals("Squid")) {
+                    databas.insert("INSERT INTO squid VALUES (" + valtID + ", " + attribut + ")");
+                }
+                JOptionPane.showMessageDialog(null, "Aliens ras och attribut har uppdaterats");
 
-        } catch (InfException ex) {
-            Logger.getLogger(AgentAndraRasPaAlien.class.getName()).log(Level.SEVERE, null, ex);
+            } catch (InfException ex) {
+                Logger.getLogger(AgentAndraRasPaAlien.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
-
     }//GEN-LAST:event_btnBekraftaActionPerformed
 
     private void cbValjAlienIDActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_cbValjAlienIDActionPerformed
@@ -277,6 +285,7 @@ public class AgentAndraRasPaAlien extends javax.swing.JFrame {
     private javax.swing.JLabel lblAndraRas;
     private javax.swing.JLabel lblNyRas;
     private javax.swing.JLabel lblRasAttribut;
+    private javax.swing.JLabel lblWorm;
     private javax.swing.JTextArea txtAreaVisaRas;
     private javax.swing.JScrollPane txtAreaVisaRasInfo;
     private javax.swing.JTextField txtRasAttribut;
